Started, but aborted.
Because it won't be able to scan the whole binary tree unless the tree
is either totally left-deep or totally right-deep. That is, we can't
guarantee that we can traverse the full tree made of just ANDs if it is
bushy. Revert to trying the other method I left off in the middle.