Add Git commit-id to the version string.
Since the git command is executed during configure step, every build
will contain the most current commit-id.

TL;DR: Use this technique to define variables in Autoconf-generated
configure files, based on shell commands.

A need for Git commit-id in Postgres version() string was expressed in
EDB-internal chat room. And a little bit of research lead me to [1].
The OP in [1] also seemed to be trying to solve a similar problem, and
there are some responses provided there, but those suggestions were
unable to realize that the configure script is processed by /bin/sh,
which is capable of embedding command-output in double-quoted strings.

After researching a bit, when I started working on Postgres'
configure.in I noticed that Postges already uses this facility, in the
version string itself, to calculate the machine-word size. So I just
extended the idea to execute Git command during configure step and get
the Git commit-id.

Obviously, this can be used to get some more descriptive infomation like
git-describe or anything much more esoteric.

This was tested with Autoconf version 2.68, and not 2.63 as approved by
Postgres (n configure.in), hence the huge amount of unrelated changes in
the configure file in this patch.

This patch is tested to work under VPATH builds (because of the `cd
$srcdir` commamd in the patch.

[1] http://stackoverflow.com/questions/8215785/how-can-i-use-ac-revision-with-git